move google calendar load to namespace
